requirements.txt
Flask
pandas
openpyxl
pdfplumber
Pillow
pytesseract
